Separate C game address locator definitions
