+++ SlashGaming-Diablo-II-API/include/c/game_address/game_offset.h
@@ -35,15 +35,46 @@
  *  work.
  */
 
-#ifndef SGD2MAPI_C_GAME_ADDRESS_H_
-#define SGD2MAPI_C_GAME_ADDRESS_H_
+#ifndef SGD2MAPI_C_GAME_ADDRESS_GAME_OFFSET_H_
+#define SGD2MAPI_C_GAME_ADDRESS_GAME_OFFSET_H_
 
-#include "game_address/game_decorated_name.h"
-#include "game_address/game_offset.h"
-#include "game_address/game_ordinal.h"
+#include "../../dllexport_define.inc"
 
-struct SGD2MAPI_GameAddress {
-  intptr_t raw_address;
-};
+#include <stdint.h>
 
-#endif // SGD2MAPI_C_GAME_ADDRESS_H_
+#ifdef __cplusplus
+extern "C" {
+#endif // __cplusplus
+
+struct SGD2MAPI_GameAddress;
+
+/**
+ * Initializes a GameAddress. The base library is specified using its ID. The
+ * game address locator is specified as the offset from the module base
+ * address to the target address.
+ */
+DLLEXPORT void
+SGD2MAPI_GameAddress_InitFromLibraryIdAndOffset(
+    struct SGD2MAPI_GameAddress* game_address,
+    int library_id,
+    intptr_t offset
+);
+
+/**
+ * Initializes a GameAddress. The base library is specified using its name,
+ * encoded in UTF-8. The game address locator is specified as the offset from
+ * the module base address to the target address.
+ */
+DLLEXPORT void
+SGD2MAPI_GameAddress_InitFromLibraryPathAndOffset(
+    struct SGD2MAPI_GameAddress* game_address,
+    const char library_path[],
+    intptr_t offset
+);
+
+#ifdef __cplusplus
+} // extern "C"
+#endif // __cplusplus
+
+#include "../../dllexport_undefine.inc"
+#endif // SGD2MAPI_C_GAME_ADDRESS_GAME_OFFSET_H_
