+++ SlashGaming-Diablo-II-API/include/c/game_address/game_ordinal.h
@@ -35,15 +35,43 @@
  *  work.
  */
 
-#ifndef SGD2MAPI_C_GAME_ADDRESS_H_
-#define SGD2MAPI_C_GAME_ADDRESS_H_
+#ifndef SGD2MAPI_C_GAME_ADDRESS_GAME_ORDINAL_H_
+#define SGD2MAPI_C_GAME_ADDRESS_GAME_ORDINAL_H_
 
-#include "game_address/game_decorated_name.h"
-#include "game_address/game_offset.h"
-#include "game_address/game_ordinal.h"
+#include "../../dllexport_define.inc"
 
-struct SGD2MAPI_GameAddress {
-  intptr_t raw_address;
-};
+#ifdef __cplusplus
+extern "C" {
+#endif // __cplusplus
 
-#endif // SGD2MAPI_C_GAME_ADDRESS_H_
+struct SGD2MAPI_GameAddress;
+
+/**
+ * Initializes a GameAddress. The base library is specified using its ID. The
+ * game address locator is specified as the address's ordinal value.
+ */
+DLLEXPORT void
+SGD2MAPI_GameAddress_InitFromLibraryIdAndOrdinal(
+    struct SGD2MAPI_GameAddress* game_address,
+    int library_id,
+    int ordinal
+);
+
+/**
+ * Initializes a GameAddress. The base library is specified using its name,
+ * encoded in UTF-8. The game address locator is specified as the address's
+ * ordinal value.
+ */
+DLLEXPORT void
+SGD2MAPI_GameAddress_InitFromLibraryPathAndOrdinal(
+    struct SGD2MAPI_GameAddress* game_address,
+    const char library_path[],
+    int ordinal
+);
+
+#ifdef __cplusplus
+} // extern "C"
+#endif // __cplusplus
+
+#include "../../dllexport_undefine.inc"
+#endif // SGD2MAPI_C_GAME_ADDRESS_GAME_ORDINAL_H_
